src/gateway/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
httpx==0.27.2
python-jose[cryptography]==3.3.0
pydantic-settings==2.4.0
tenacity==9.0.0
prometheus-client==0.20.0
